Extracts app delegate into separate file.
