+++ Bolistik/Bolistik/App/AppManager.swift
